requirements.txt
pandas
numpy
scikit-learn
mlflow
pytest
black
lint
flake8
isort
requests
seaborn
xgboost
lightgbm
shap
fastapi
pyyaml